requirements.txt
django
djangorestframework
requests
pyyaml
django-cors-headers